Move Tree test struct to worked `RecursionlessSizeBalancedTree`
