+++ rust/src/methods/trees/size_balanced_tree.rs
@@ -263,118 +263,3 @@ pub trait SizeBalancedTreeMethods<T: Num> {
         unsafe { self.detach_core(root, node) }
     }
 }
-
-#[derive(Default, Debug, Clone, Copy)]
-pub struct TreeElement<T: Num> {
-    pub size: T,
-    pub left: T,
-    pub right: T,
-}
-
-const N: usize = 10000;
-
-pub struct SizeBalancedTree<T: Num> {
-    pub elements: [TreeElement<T>; N],
-    _allocated: T,
-    pub root: T,
-}
-
-impl<T: Num> SizeBalancedTree<T> {
-    pub fn new() -> Self {
-        Self {
-            elements: [Default::default(); N],
-            _allocated: T::one(),
-            root: T::zero(),
-        }
-    }
-
-    pub fn alloc(&mut self) -> T {
-        let new = self._allocated;
-        if self.is_empty(new) {
-            self._allocated = self._allocated + T::one();
-            return new;
-        }
-        // TODO: return error
-        panic!("allocated tree element is not empty");
-    }
-
-    pub fn dealloc(&mut self, mut node: T) {
-        while self._allocated != T::one() && self.is_empty(node) {
-            let last = self._allocated - T::one();
-            if last == node {
-                self._allocated = last;
-                node = node - T::one();
-            } else {
-                return;
-            }
-        }
-    }
-
-    pub fn len(&self) -> T {
-        self.get_size(self.root)
-    }
-
-    pub fn is_empty(&self, node: T) -> bool {
-        unsafe fn is_zero(ptr: *const usize, mut bytes: isize) -> bool {
-            let ptr = ptr as *const i8;
-            (0..bytes).all(|i| *(ptr.offset(i)) == 0)
-        }
-
-        unsafe { is_zero(self.get(node) as *mut TreeElement<T> as *const usize, size_of::<TreeElement<T>>() as isize) }
-    }
-
-    fn get(&self, node: T) -> &mut TreeElement<T> {
-        // TODO: lol stupid implementation
-        unsafe { &mut *(&self.elements[node.as_()] as *const TreeElement<T> as *mut TreeElement<T>) }
-    }
-}
-
-impl<T: Num> SizeBalancedTreeMethods<T> for SizeBalancedTree<T> {
-    fn get_left_reference(&self, node: T) -> *const T {
-        &self.get(node).left as *const T
-    }
-
-    fn get_right_reference(&self, node: T) -> *const T {
-        &self.get(node).right as *const T
-    }
-
-    fn get_mut_left_reference(&mut self, node: T) -> *mut T {
-        &mut self.get(node).left as *mut T
-    }
-
-    fn get_mut_right_reference(&mut self, node: T) -> *mut T {
-        &mut self.get(node).right as *mut T
-    }
-
-    fn get_left(&self, node: T) -> T {
-        self.get(node).left
-    }
-
-    fn get_right(&self, node: T) -> T {
-        self.get(node).right
-    }
-
-    fn get_size(&self, node: T) -> T {
-        self.get(node).size
-    }
-
-    fn set_left(&mut self, node: T, left: T) {
-        self.get(node).left = left
-    }
-
-    fn set_right(&mut self, node: T, right: T) {
-        self.get(node).right = right
-    }
-
-    fn set_size(&mut self, node: T, size: T) {
-        self.get(node).size = size
-    }
-
-    fn first_is_to_the_left_of_second(&self, first: T, second: T) -> bool {
-        first < second
-    }
-
-    fn first_is_to_the_right_of_second(&self, first: T, second: T) -> bool {
-        first > second
-    }
-}
